Improvements for PR #193 (Step and Row Data inclusion)
* changes type of total cycles from Short to Long (does not fit for cycling)
* enhanced unit tests for parsing total cycles from FIT files
* don't pass totalCycles to c'tor when not available
* text resource improvements
* added to Changelog

+++ st-exerciseviewer/src/main/kotlin/de/saring/exerciseviewer/data/ExerciseCadence.kt
@@ -3,9 +3,9 @@ package de.saring.exerciseviewer.data
 /**
  * This class stores the cadence summary of a recorded exercise.
  *
- * @property cadenceAvg Average cadence of exercise (rpm).
- * @property cadenceMax Maximum cadence of exercise (rpm).
- * @property cadenceTotal Total cycles of exercise recorded.
+ * @property cadenceAvg Average cadence of exercise (rpm)
+ * @property cadenceMax Maximum cadence of exercise (rpm)
+ * @property cadenceTotal Total cycles of exercise recorded (optional).
  *
  * @author Stefan Saring
  */
@@ -13,4 +13,4 @@ data class ExerciseCadence(
 
     var cadenceAvg: Short,
     var cadenceMax: Short,
-    var cyclesTotal: Short? = null)
+    var cyclesTotal: Long? = null)
